Share segments_of_year fixture
